Payment, one-to-one association

+++ src/main/java/com/StudyProject/course/entities/Product.java
@@ -25,7 +25,7 @@ public class Product implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String description;
 	private Double price;
@@ -38,7 +38,7 @@ public class Product implements Serializable {
 	public Product() {
 	}
 
-	public Product(Integer id, String name, String description, Double price, String imgUrl) {
+	public Product(Long id, String name, String description, Double price, String imgUrl) {
 		this.id = id;
 		this.name = name;
 		this.description = description;
@@ -49,11 +49,11 @@ public class Product implements Serializable {
 	@OneToMany (mappedBy = "id.product") 	
 	private Set<OrderItem> items = new HashSet<>();
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
